simply dictionary format, improve internal naming
Change dictionary format for text_after for a command into a string that
is parsed into text actions, instead of asking the user to write the
serialized text actions.

Change internal Attached.do_orthography name and type (from Option<bool>
to an enum) so the name better conveys its purpose.

Added an option for a custom config folder in the command line options.